+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 0.5.3 (2025-10-17)
+
+### Features
+
+- Updated System.CommandLine dependency to 2.5 beta
+
+### Bugs Fixed
+
+- Added a language specific way to get package name for validation checks, to account for different language naming (JS uses package.json name)
+
 ## 0.5.2 (2025-10-13)
 
 ### Features